backend/src/movies/api/serializers.py
from rest_framework import serializers
from movies.models import Article

class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = ('id', 'title', 'content')
