backend/requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
pytest==4.4.0
pytest-mock==1.10.3
pytest-watch==4.2.0
redis==3.3.11
tmdbv3api==1.5.0
django-cors-headers==2.4.0